Modify(junsang/board): 게시글, 동아리 게시판 게시글 api 리덕스 적용

+++ src/apis/post.js
@@ -2,12 +2,21 @@ import axios from 'apis/index';
 
 const api = {
   getPost: (category, pid) => {
-    return axios.get(`/api/board/${category}/${pid}`)
+    if (category === 'clubNotice') {
+      return axios.get(`/api/club/board/clubNotice/2/${pid}`);
+    }
+    return axios.get(`/api/board/${category}/${pid}`);
   },
   deletePost: (category, pid) => {
-    return axios.delete(`/api/board/${category}/${pid}`)
+    if (category === 'clubNotice') {
+      return axios.get(`/api/club/board/clubNotice/2/${pid}`);
+    }
+    return axios.delete(`/api/board/${category}/${pid}`);
   },
   postComment: ({ category, pid, id, description, parentCommentID }) => {
+    if (category === 'clubNotice') {
+      return axios.post(`/api/club/board/clubNotice/2/${pid}`, { id, description });
+    }
     if (parentCommentID) {
       return axios.post(`/api/board/${category}/${pid}/${parentCommentID}`, { id, description });
     } else {
@@ -15,6 +24,9 @@ const api = {
     }
   },
   putComment: ({ category, pid, commentID, description, parentCommentID }) => {
+    if (category === 'clubNotice') {
+      return axios.put(`/api/club/board/clubNotice/2/${pid}/${commentID}`, { description });
+    }
     if (parentCommentID) {
       return axios.put(`/api/board/${category}/${pid}/${parentCommentID}/${commentID}`, { description });
     } else {
@@ -22,6 +34,9 @@ const api = {
     }
   },
   deleteComment: ({ category, pid, commentID, parentCommentID }) => {
+    if (category === 'clubNotice') {
+      return axios.delete(`/api/club/board/clubNotice/2/${pid}/${commentID}`);
+    }
     if (parentCommentID) {
       return axios.delete(`/api/board/${category}/${pid}/${parentCommentID}/${commentID}`);
     } else {
